+++ src/components/common/ThumbnailGrid/Thumbnail/Thumbnail.stories.tsx
