+++ src/components/common/ThumbnailGrid/ThumbnailGrid.stories.tsx
