+++ src/components/common/ThumbnailGrid/index.ts
@@ -1,0 +1,3 @@
+export { default as ThumbnailGrid } from "./ThumbnailGrid";
+export { default as Thumbnail } from "./Thumbnail/Thumbnail";
+export * from "./skeleton";
