+++ src/components/common/ThumbnailGrid/skeleton/ThumbnailSkeleton.tsx
@@ -1,0 +1,7 @@
+import { Skeleton } from "@/components/ui/skeleton";
+
+const ThumbnailSkeleton = () => {
+  return <Skeleton className="absolute left-0 top-0 h-full w-full rounded" />;
+};
+
+export default ThumbnailSkeleton;
