+++ src/components/common/ThumbnailGrid/skeleton/index.ts
@@ -1,0 +1,2 @@
+export { default as ThumbnailSkeleton } from "./ThumbnailSkeleton";
+export { default as ThumbnailGridSkeleton } from "./ThumbnailGridSkeleton";
