--- src/components/common/thumbnail/index.ts
@@ -1,2 +1,0 @@
-export { default as Thumbnail } from "./Thumbnail";
-export { default as ThumbnailGrid } from "./ThumbnailGrid";
